+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - switch from `async-std` to `tokio` as the async runtime #3449
 - upgrade to `pgp@0.8.0` #3467
 - add IMAP ID extension support #3468
+- configure DeltaChat folder by selecting it, so it is configured even if not LISTed #3371
 
 ### Fixes
 - mailing list: remove square-brackets only for first name #3452